+++ tests/bootstrap.php
@@ -39,6 +39,7 @@ if (!Plugin::isPluginActive("advancedforms")) {
     throw new RuntimeException("Plugin advancedforms is not active in the test database");
 }
 
+require __DIR__ . "/ConfigurableItemsTrait.php";
 require __DIR__ . "/AdvancedFormsTestCase.php";
 require __DIR__ . "/Front/FrontTestCase.php";
 require __DIR__ . "/Model/Mapper/MapperTestCase.php";